+++ src/hooks/useInitialTradesData.ts
